TrackSelection: move back to boolean columns

+++ Analysis/Tasks/trackselection.cxx
@@ -108,8 +108,8 @@ struct TrackSelectionTask {
   void process(soa::Join<aod::FullTracks, aod::TracksExtended> const& tracks)
   {
     for (auto& track : tracks) {
-      filterTable((uint8_t)globalTracks.IsSelected(track),
-                  (uint8_t)globalTracksSDD.IsSelected(track));
+      filterTable(globalTracks.IsSelected(track),
+                  globalTracksSDD.IsSelected(track));
     }
   }
 };
